+++ lib/scancommands.dart
@@ -1,10 +1,10 @@
-import 'package:scanwedge/scanwedge_channel.dart';
-
 class ScanCommands {
+  static const datawedgeSendSetConfig = 'com.symbol.datawedge.api.SET_CONFIG';
+  static const profileIntentAction = 'no.talgoe.scanwedge.scanwedge.SCAN';
   static const scannerInputPlugin = 'com.symbol.datawedge.api.SCANNER_INPUT_PLUGIN';
-  static Future<bool> enableScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'ENABLE_PLUGIN');
-  static Future<bool> disableScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'DISABLE_PLUGIN');
-  static Future<bool> suspendScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'SUSPEND_PLUGIN');
-  static Future<bool> resumeScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'RESUME_PLUGIN');
-  static Future<bool> toggleScanning() => ScanwedgeChannel.instance.toggleScanning();
+  // static Future<bool> enableScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'ENABLE_PLUGIN');
+  // static Future<bool> disableScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'DISABLE_PLUGIN');
+  // static Future<bool> suspendScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'SUSPEND_PLUGIN');
+  // static Future<bool> resumeScanner() => ScanwedgeChannel.instance.sendCommand(command: scannerInputPlugin, parameter: 'RESUME_PLUGIN');
+  // static Future<bool> toggleScanning() => ScanwedgeChannel.instance.toggleScanning();
 }