+++ lib/scanwedge_channel.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:developer';
+import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
@@ -7,20 +8,49 @@ import 'package:scanwedge/models/scanresult.dart';
 
 class ScanwedgeChannel {
   static const channel = 'scanwedge';
-  static final ScanwedgeChannel instance = ScanwedgeChannel._init();
-  final MethodChannel _methodChannel = const MethodChannel(channel);
-  final _streamController = StreamController<ScanResult>();
-  Stream<ScanResult> get stream => _streamController.stream;
-  ScanwedgeChannel._init() {
+  // static final ScanwedgeChannel instance = ScanwedgeChannel._init();
+  static const _methodChannel = MethodChannel(channel);
+  final _streamController = StreamController<ScanResult>.broadcast();
+  bool isZebra = false;
+  //"${android.os.Build.MANUFACTURER}|${android.os.Build.MODEL}|${android.os.Build.PRODUCT}|${android.os.Build.VERSION.RELEASE}"
+  String? _manufacturer, _model, _product, _osVersion, _packageName;
+  bool get isDeviceSupported => isZebra;
+  String get modelName => _model ?? '';
+  String get productName => _product ?? '';
+  String get manufacturer => _manufacturer ?? '';
+  String get osVersion => _osVersion ?? '';
+  String get packageName => _packageName ?? '';
+  ScanwedgeChannel._({required String? deviceInfo}) {
     _methodChannel.setMethodCallHandler(_methodHandler);
+    // getDeviceInfo().then((deviceInfo) {
+    debugPrint('getDeviceInfo: $deviceInfo');
+    if (deviceInfo != null) {
+      final devInfoString = deviceInfo.split('|');
+      _manufacturer = devInfoString.first;
+      if (deviceInfo.length > 3) {
+        _model = devInfoString[1];
+        _product = devInfoString[2];
+        _osVersion = devInfoString[3];
+        isZebra = _manufacturer!.toUpperCase().startsWith('ZEBRA') || _model!.toUpperCase().startsWith('ZEBRA');
+        debugPrint('deviceInfo($_manufacturer, $_model, $_product, $_osVersion)-$isZebra');
+      }
+    }
+    // });
   }
+  Stream<ScanResult> get stream => _streamController.stream;
+  static Future<ScanwedgeChannel> initialize() async {
+    debugPrint('init called');
+    if (!kIsWeb && Platform.isAndroid) {
+      final deviceInfo = await _methodChannel.invokeMethod<String>('getDeviceInfo');
+      return ScanwedgeChannel._(deviceInfo: deviceInfo);
+    }
+    return ScanwedgeChannel._(deviceInfo: null);
+  }
+
   Future<void> _methodHandler(MethodCall call) async {
     debugPrint("_methodHandler($call, ${call.arguments})-${call.method}");
     log("_methodHandler($call, ${call.arguments})");
     switch (call.method) {
-      case "test":
-        log("test kalt: ${call.arguments}");
-        break;
       case "scan":
         debugPrint("scan method ${call.arguments.runtimeType}");
         final scanResult = ScanResult.fromDatawedge(call.arguments);
@@ -33,14 +63,15 @@ class ScanwedgeChannel {
     }
   }
 
-  Future<String?> getPlatformVersion() async => await _methodChannel.invokeMethod<String>('getPlatformVersion');
+  Future<String?> getDeviceInfo() async => await _methodChannel.invokeMethod<String>('getDeviceInfo');
 
-  Future<bool> createProfile({required String profileName, required String packageName}) async =>
-      await _methodChannel.invokeMethod<bool>('createProfile', {'profileName': profileName, 'packageName': packageName}) ?? false;
+  Future<bool> toggleScanning() async => isDeviceSupported ? await _methodChannel.invokeMethod<bool>('toggleScan') ?? false : false;
 
-  Future<bool> toggleScanning() async => await _methodChannel.invokeMethod<bool>('toggleScan') ?? false;
+  Future<bool> sendCommand({required String command, required String parameter}) async =>
+      isDeviceSupported ? await _methodChannel.invokeMethod<bool>('sendCommand', {'command': command, 'parameter': parameter}) ?? false : false;
 
-  Future<bool> sendCommand({required String command, required String parameter}) async => await _methodChannel.invokeMethod<bool>('sendCommand', {'command': command, 'parameter': parameter}) ?? false;
-  Future<bool> sendCommandBundle({required String command, required Map<String, dynamic> parameter, required bool sendResult}) async =>
-      await _methodChannel.invokeMethod<bool>('sendCommandBundle', {'command': command, 'parameter': parameter, 'sendResult': sendResult}) ?? false;
+  Future<bool> sendCommandBundle({required String command, required Map<String, dynamic> parameter, required bool sendResult}) async {
+    debugPrint('sendCommandBundle-$isDeviceSupported');
+    return isDeviceSupported ? await _methodChannel.invokeMethod<bool>('sendCommandBundle', {'command': command, 'parameter': parameter, 'sendResult': sendResult}) ?? false : false;
+  }
 }